Check all routes for continuity before generating Javascript. The fix carries two additional benefits:  it fixes display anomalies where paths are shared among multiple layers and also ensures the reuse of common data when generating Javascript files--for a 75% reduction in size.
This operation uses the current route-checking capability to
reconstruct the contents of routes and segments.  While it
aims to achieve maximal reusability of segments and paths
within routes, it requires care around split points.  These
indicate the reuse of only parts of segments or paths.  In
such cases, the code generates new segments and paths to
represent the partial data.  Such newly generated segments
and paths may themselves be reused by other routes defined
in terms of them.

The program now includes the -relax-route-check command-line
switch to disable this checking.  This feature may be useful
during the development of the source dataset by letting the
user visualize how to resolve a noncontinuous route in the
data.  This feature is not for production builds.

Added and modified unit tests to reflect the new feature
both with and without the new -relax-route-check flag
applied.  Added a note to the man page for the new switch.